Move all client methods to client.go; move all constants to constants.go
