+++ device.go
@@ -187,82 +187,3 @@ func (d *Device) SubscribeNotifications(names []string, timestamp time.Time) (su
 
 	return d.client.SubscribeNotifications(params)
 }
-
-func (c *Client) GetDevice(deviceId string) (device *Device, err *Error) {
-	device = &Device{
-		client: c,
-	}
-
-	err = c.getModel("getDevice", device, map[string]interface{}{
-		"deviceId": deviceId,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return device, nil
-}
-
-func (c *Client) PutDevice(deviceId, name string, data map[string]interface{}, networkId, deviceTypeId int64, isBlocked bool) (device *Device, err *Error) {
-	device = &Device{
-		client: c,
-	}
-
-	device.Id = deviceId
-
-	if name == "" {
-		device.Name = deviceId
-	} else {
-		device.Name = name
-	}
-
-	if data != nil {
-		device.Data = data
-	}
-
-	if networkId != 0 {
-		device.NetworkId = networkId
-	}
-
-	if deviceTypeId != 0 {
-		device.DeviceTypeId = deviceTypeId
-	}
-
-	if isBlocked {
-		device.IsBlocked = isBlocked
-	}
-
-	_, err = c.request("putDevice", map[string]interface{}{
-		"deviceId": deviceId,
-		"device":   device,
-	})
-
-	if err != nil {
-		return nil, err
-	}
-
-	return device, nil
-}
-
-func (c *Client) ListDevices(params *ListParams) (list []*Device, err *Error) {
-	if params == nil {
-		params = &ListParams{}
-	}
-
-	data, pErr := params.Map()
-	if pErr != nil {
-		return nil, &Error{name: InvalidRequestErr, reason: pErr.Error()}
-	}
-
-	rawRes, err := c.request("listDevices", data)
-	if err != nil {
-		return nil, err
-	}
-
-	pErr = json.Unmarshal(rawRes, &list)
-	if pErr != nil {
-		return nil, newJSONErr()
-	}
-
-	return list, nil
-}
