+++ error.go
@@ -2,13 +2,6 @@ package devicehive_go
 
 import "github.com/devicehive/devicehive-go/internal/transport"
 
-const (
-	InvalidResponseErr  = "invalid response"
-	InvalidRequestErr   = "invalid request"
-	ServiceErr          = "service error"
-	ConnectionFailedErr = "connection failed"
-)
-
 func newError(err error) *Error {
 	if err == nil {
 		return nil
